Add more components and reconfigure for back-end catch-all routing
As we have front-end routing setup for index.html, we want back-end to simply provide index.html whenever a page is requested.
This solution is far from perfect, as we have to manually add new pages to server.js as we add more. This solution should be replaced with an isophomorphic approach later.